+++ src/aosm/azext_aosm/deploy/deploy_with_arm.py
@@ -11,13 +11,13 @@ from typing import Any, Dict, Optional
 
 from knack.log import get_logger
 from azext_aosm.deploy.artifact_manifest import ArtifactManifestOperator
-from azext_aosm.util.management_clients import ApiClientsAndCaches
+from azext_aosm.util.management_clients import ApiClients
 from azure.mgmt.resource.resources.v2021_04_01.models import DeploymentExtended
 from pathlib import Path
 
 from azext_aosm.deploy.pre_deploy import PreDeployerViaSDK
-from azext_aosm._configuration import Configuration, VNFConfiguration
-from azext_aosm._constants import (
+from azext_aosm.configuration import Configuration, VNFConfiguration
+from azext_aosm.util.constants import (
     VNF_DEFINITION_BICEP_SOURCE_TEMPLATE,
     VNF_MANIFEST_BICEP_SOURCE_TEMPLATE,
 )
@@ -33,7 +33,7 @@ class DeployerViaArm:
     # using the SDK
     def __init__(
         self,
-        apiClientsAndCaches: ApiClientsAndCaches,
+        api_clients: ApiClients,
         config: Configuration,
     ) -> None:
         """
@@ -45,9 +45,9 @@ class DeployerViaArm:
         :type resource_client: ResourceManagementClient
         """
         logger.debug("Create ARM/Bicep Deployer")
-        self.api_clients = apiClientsAndCaches
+        self.api_clients = api_clients
         self.config = config
-        self.pre_deployer = PreDeployerViaSDK(apiClientsAndCaches, self.config)
+        self.pre_deployer = PreDeployerViaSDK(api_clients, self.config)
 
     def deploy_vnfd_from_bicep(
         self,
