+++ src/aosm/azext_aosm/vendored_sdks/_hybrid_network_management_client.py
@@ -14,7 +14,7 @@ from msrest import Deserializer, Serializer
 from azure.mgmt.core import ARMPipelineClient
 
 from . import models
-from ._configuration import HybridNetworkManagementClientConfiguration
+from configuration import HybridNetworkManagementClientConfiguration
 from .operations import ArtifactManifestsOperations, ArtifactStoresOperations, ComponentsOperations, ConfigurationGroupSchemasOperations, ConfigurationGroupValuesOperations, HybridNetworkManagementClientOperationsMixin, NetworkFunctionDefinitionGroupsOperations, NetworkFunctionDefinitionVersionsOperations, NetworkFunctionReadyK8SOperations, NetworkFunctionsOperations, NetworkServiceDesignGroupsOperations, NetworkServiceDesignVersionsOperations, Operations, PreviewSubscriptionsOperations, ProxyArtifactOperations, ProxyNetworkFunctionDefinitionGroupsOperations, ProxyNetworkFunctionDefinitionVersionsOperations, ProxyPublisherOperations, PublishersOperations, SiteNetworkServicesOperations, SitesOperations
 
 if TYPE_CHECKING:
